+++ src/main/clojure/clojure/tools/gitlibs/impl.clj
@@ -24,15 +24,22 @@
   (binding [*out* *err*]
     (apply println msgs)))
 
-(defn capture
-  "Reads from input-stream until EOF and returns a String (or nil if 0 length).
-   Takes same opts as clojure.java.io/copy - :buffer and :encoding"
+(defn- capture
+  "Reads from input-stream until EOF and returns a String (or nil if 0 length)."
   [^InputStream input-stream]
   (let [writer (StringWriter.)]
     (jio/copy input-stream writer)
     (let [s (str/trim (.toString writer))]
       (when-not (zero? (.length s))
         s))))
+
+(defmacro background
+  [& body]
+  `(let [result# (promise)]
+     (doto (Thread. (fn [] (deliver result# (do ~@body))))
+       (.setDaemon true)
+       (.start))
+     result#))
 
 (defn- run-git
   [& args]
@@ -44,8 +51,8 @@
           _ (when debug (.redirectError proc-builder ProcessBuilder$Redirect/INHERIT))
           _ (when-not terminal (.put (.environment proc-builder) "GIT_TERMINAL_PROMPT" "0"))
           proc (.start proc-builder)
-          out (future (capture (.getInputStream proc)))
-          err (future (capture (.getErrorStream proc))) ;; if debug is true, stderr will be redirected instead
+          out (background (capture (.getInputStream proc)))
+          err (background (capture (.getErrorStream proc))) ;; if debug is true, stderr will be redirected instead
           exit (.waitFor proc)]
       {:args command-args, :exit exit, :out @out, :err @err})))